Final submission. Functional.
Had to revert from using if-not conditonal, as it lead to a process flow problem, whereby every question/decision had to be syntatically followed by a newly defined if-statement, therefore continuing the loop even if the previous outcome was GAME OVER. Also used color syntax in my print statement so just be aware of that. Otherwise all possible outcomes are covered as specified in the diagram.